Perfect UMAP and spatial map plots
The colors between graphs are now related.  
Special function to choose and modify the appropriate colormap and put 
the noise color at the start or end of the colormap.
Try several min_cluster_size values for HDBSCAN to have clusters that 
make sense, on the UMAP projection and on the spatial map.